Add storeSliceIntoStruct
Presently, if you have a struct _within_ a struct or map, godbus throws an error that []dbus.Variant cannot be mapped to a struct.

This is not true. However, the way things flow to get to that point, there's no easy way to get back to mapping a slice of variants to a struct.

So we instead add a storeSliceIntoStruct() call and everything is happy.